Rename CloudSolrClient to CloudHttp1SolrClient Rename BaseCloudSolrClient to CloudSolrClient
